Everything: rewrite the whole thing to not use deb tools, but fpm
Long story short, this commit[1] was supposedly the way to go, but it
didn't really work because it caused this[2]. Now, apparently we were
using dpkg-buildpackage in a slightly incorrect way because this tool
can in theory call ./configure && make itself, but I haven't managed to
figure out how. Fuck debian and its complexity. (Read about 20 tutorials
and still don't have a clue how to do it; if you know how, answer this[3].)

fpm is a much saner option, because we're not trying to contribute all this
to debian upstream. Just read fpm end goal here[4].

Oh BTW, this commit also removes our custom patch to apply to 3.0.3, because
nowadays we can just use 3.0.4 or 3.0.6 (the latter is a bugfix release to
the former).

[1] https://github.com/7digital/7digital-mono/commit/e6fb5b98d089836393ef57dea898eeaf3a64ef06
[2] http://stackoverflow.com/questions/15187344/dpkg-buildpackage-uc-us-b-fails-with-dpkg-shlibdeps-error-couldnt-find-li
[3] http://stackoverflow.com/questions/15231016/simplest-way-of-using-dpkg-buildpackage-with-an-existing-upstream-source-tarball
[4] https://github.com/jordansissel/fpm/blob/master/README.md